Add files via upload
Source code is provided in BDD_Project2.2.docx
For problem 2, UDF is written to compare the same date and code is written in file DateComp.java. Its jar file is checkSameDate.jar

Screen shot is provided for data stored in HDFS, also screenshot is provided for the output on console screen using DUMP command.

After exporting the data from HDFS to MySql using Sqoop command, screenshot is provided for data in Mysql.